+++ hutool-core/src/test/java/com/xiaoleilu/hutool/util/RuntimeUtilTest.java
@@ -5,12 +5,24 @@ import org.junit.Test;
 
 import com.xiaoleilu.hutool.lang.Console;
 
+/**
+ * 命令行单元测试
+ * @author looly
+ *
+ */
 public class RuntimeUtilTest {
 
 	@Test
 	@Ignore
 	public void execTest() {
 		String str = RuntimeUtil.execForStr("ipconfig");
+		Console.log(str);
+	}
+
+	@Test
+	@Ignore
+	public void execCmdTest() {
+		String str = RuntimeUtil.execForStr("cmd /c dir");
 		Console.log(str);
 	}
 }